+++ changelog.md
@@ -1,10 +1,41 @@
 # Changelog
 For Techy People
 
-## 1.8.1
+# Version 1.9 - Sorting out your internals, Hexfriend
+
+## Version 1.9.0
+Tile names now local scoped to tileset ID
+Texture loader revamped to identify tile textures based on tileset ID and tile name
+Save data updated to v10
+
+# Version 1.8 - New Stripes, Hexfriend
+
+## Version 1.8.6
+Clicking left mouse + right mouse at the same time no longer makes panning sticky
+Releasing right click over the toolbar no longer makes panning stick
+
+## Version 1.8.5
+Icon scale UX improvements
+Icon Eyedropper
+Save data version update
+Note: due to code choices the icon eyedropper will not pick out size on old maps
+
+
+## Version 1.8.4
+- Coordinates bug fix
+- Save data converter path fix
+
+## Version 1.8.3
+- Icon scale bug fix
+
+## Version 1.8.2
+- Grid gap added (thanks Evan!)
+- Save data update to version 8
+
+## Version 1.8.1
 - Added changelog
 
-## 1.8.0
+## Version 1.8.0
 - Added dashed lines
 - Path data change (paths now have dash info in style)
 - Path data moved to store